Implements #243
As there are significant differences in wheel rotation, the approach now
has been to create a set of markers for {low, normal, high} range wheels
which allows all of them to be adapted independently.

All ranges start at 90, which is the lowest value, so the Scale will
always have a leftmost value. Tested with all current values in
device_manager.py.